[#12159] Store elapsed time directly as a column value instead of using a variable
